backend/MLmodel/requirements.txt
tensorflow==2.15.0
scikeras==0.12.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
joblib==1.3.2
scipy==1.13.0
yfinance==0.2.65
